feat: 공통 모듈 생성
